don't set tree, clusters, is_leaf as np.float32, np.float64 since these will only ever be signed integers
sort top weights

+++ hNMF/model.py
@@ -246,7 +246,7 @@ class HierarchicalNMF(BaseEstimator):
             x = np.zeros(self.n_features_)
             x[cluster] = 1
             stacked.append(x)
-        return np.vstack(stacked)
+        return np.vstack(stacked).astype(np.int)
 
     def _stack_H_buffer(self, buffer: list) -> np.ndarray:
         # Returns components_ with shape (2*k-1, 2, n_features)
